Move sim files to sim directory
