+++ packages/cli/test/sim/errors.ts
