+++ packages/cli/test/sim/eth2clients/index.ts
