+++ packages/cli/test/sim/eth2clients/interface.ts
