+++ packages/cli/test/sim/participation.ts
