+++ packages/cli/test/sim/utils/math.ts
