+++ packages/cli/test/sim/utils/tableRender.ts
