+++ packages/cli/test/sim/utils/timeLogGenesis.ts
